+++ lib/dot.dart
@@ -1,21 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/physics.dart';
 
+import 'shape.dart';
+
+const defaultActiveDotColor = Color(0xff35affc);
+
 class DotInstance extends StatefulWidget {
   DotInstance({
     Key key,
     @required this.listenable,
     @required this.length,
-    this.size = 16,
-    this.spacing = 8,
-    this.color = const Color(0xff35affc),
+    this.shape,
+    this.color = defaultActiveDotColor,
   }) : super(key: key);
 
   final PageController listenable;
-  final int size;
-  final int spacing;
   final int length;
   final Color color;
+  final Shape shape;
   @override
   State<StatefulWidget> createState() => DotInstanceState();
 }
@@ -26,13 +28,15 @@ class DotInstanceState extends State<DotInstance>
   double _page = 0;
   final ceilRange = 0.999999;
   final floorRange = 0.0001;
-  SpringDescription spring = SpringDescription(mass: 1.0, stiffness: 100.0, damping: 10.0);
+  SpringDescription spring = SpringDescription(
+    mass: 1.0,
+    stiffness: 100.0,
+    damping: 10.0,
+  );
   SpringSimulation springSimulation;
   AnimationController animationController;
 
-  @override
-  void initState() {
-    super.initState();
+  void setUpWidgetListenable() {
     widget.listenable.addListener(() {
       setState(() {
         _offset = widget.listenable.page;
@@ -41,7 +45,9 @@ class DotInstanceState extends State<DotInstance>
         }
       });
     });
+  }
 
+  void setUpAnimationController() {
     animationController = new AnimationController(
       vsync: this,
       lowerBound: double.negativeInfinity,
@@ -49,14 +55,50 @@ class DotInstanceState extends State<DotInstance>
     );
   }
 
+  @override
+  void initState() {
+    super.initState();
+    setUpWidgetListenable();
+    setUpAnimationController();
+  }
+
   double getMargin(context, length) {
     double width = MediaQuery.of(context).size.width;
-    var leftMargin = (width - length * widget.size - (length - 1) * widget.spacing) / 2;
+    var leftMargin = (width -
+            length * widget.shape.width -
+            (length - 1) * widget.shape.spacing) /
+        2;
     if (_offset >= _page) {
-      return leftMargin + (widget.size + widget.spacing) * _page;
+      return leftMargin + (widget.shape.width + widget.shape.spacing) * _page;
     }
-    else {
-      return leftMargin + (widget.size + widget.spacing) * _offset;
+
+    return leftMargin + (widget.shape.width + widget.shape.spacing) * _offset;
+  }
+
+  BoxDecoration _getBoxDecoration(Shape dotShape) {
+    switch (dotShape.shape) {
+      case DotShape.Circle:
+        return BoxDecoration(
+          color: widget.color ?? defaultActiveDotColor,
+          shape: BoxShape.rectangle,
+          borderRadius: BorderRadius.circular(
+            widget.shape.size / 2,
+          ),
+        );
+      case DotShape.Rectangle:
+      case DotShape.Square:
+        return BoxDecoration(
+          color: widget.color ?? defaultActiveDotColor,
+          shape: BoxShape.rectangle,
+        );
+      default:
+        return BoxDecoration(
+          color: widget.color ?? defaultActiveDotColor,
+          shape: BoxShape.rectangle,
+          borderRadius: BorderRadius.circular(
+            widget.shape.width / 2,
+          ),
+        );
     }
   }
 
@@ -64,17 +106,18 @@ class DotInstanceState extends State<DotInstance>
   Widget build(BuildContext context) {
     final width = (_offset - _page).abs().toDouble();
     return Container(
-      margin: EdgeInsets.only(left: getMargin(context, widget.length)),
+      margin: EdgeInsets.only(
+        left: getMargin(context, widget.length),
+      ),
       child: Container(
-        width: width <= floorRange ? widget.size.toDouble() :
-        widget.size + (widget.size + widget.spacing) * (_offset - _page).abs().toDouble(),
-        height: widget.size.toDouble(),
+        width: width <= floorRange
+            ? widget.shape.width
+            : widget.shape.width +
+                (widget.shape.width + widget.shape.spacing) *
+                    (_offset - _page).abs().toDouble(),
+        height: widget.shape.height,
       ),
-      decoration: BoxDecoration(
-        color: widget.color ?? Color(0xff35affc),
-        shape: BoxShape.rectangle,
-        borderRadius: BorderRadius.circular(widget.size / 2),
-      ),
+      decoration: _getBoxDecoration(widget.shape),
     );
   }
 }