+++ lib/worm_indicator.dart
@@ -1,119 +1,121 @@
 library worm_indicator;
 
 import 'package:flutter/material.dart';
+
 import 'dot.dart';
+import 'shape.dart';
+
+const defaultNormalDotColor = Color(0xff808080);
 
 class WormIndicator extends StatefulWidget {
   WormIndicator({
     Key key,
     @required this.length,
-    this.spacing = 8,
-    this.size = 16,
     this.controller,
+    this.shape,
     this.color = const Color(0xff808080),
     this.indicatorColor = const Color(0xff35affc),
   }) : super(key: key);
 
   final int length;
-  final size;
-  final spacing;
   final PageController controller;
   final Color color;
   final Color indicatorColor;
+  final Shape shape;
+
   @override
   State<StatefulWidget> createState() => _DotsIndicatorState();
 }
 
 class _DotsIndicatorState extends State<WormIndicator> {
+  Container getNormalDotChildContainer() {
+    switch (widget.shape.shape) {
+      case DotShape.Circle:
+        return Container(
+          width: widget.shape.size,
+          height: widget.shape.size,
+        );
+      default:
+        return Container(
+          width: widget.shape.width,
+          height: widget.shape.height,
+        );
+    }
+  }
+
+  BoxDecoration getNormalDotDecoration(Color color) {
+    switch (widget.shape.shape) {
+      case DotShape.Circle:
+        return BoxDecoration(
+          color: color ?? Color(0xff35affc),
+          shape: BoxShape.circle,
+        );
+      case DotShape.Square:
+      case DotShape.Rectangle:
+        return BoxDecoration(
+          color: color ?? Color(0xff35affc),
+          shape: BoxShape.rectangle,
+        );
+      default:
+        return BoxDecoration(
+          color: color ?? Color(0xff35affc),
+          shape: BoxShape.circle,
+        );
+    }
+  }
+
   Widget buildDot(color, index) {
     if ((widget.length % 2 == 1 && index == (widget.length ~/ 2)) ||
         index == -1) {
       return Container(
-        child: Container(
-          width: widget.size.toDouble(),
-          height: widget.size.toDouble(),
-        ),
-        decoration: BoxDecoration(
-          color: color ?? Color(0xff35affc),
-          shape: BoxShape.circle,
-        ),
+        child: getNormalDotChildContainer(),
+        decoration: getNormalDotDecoration(color),
       );
     }
 
     if (widget.length % 2 == 1 && index < (widget.length ~/ 2)) {
       return Container(
-        margin: EdgeInsets.only(right: widget.spacing.toDouble()),
-        child: Container(
-          width: widget.size.toDouble(),
-          height: widget.size.toDouble(),
-        ),
-        decoration: BoxDecoration(
-          color: color ?? Color(0xff35affc),
-          shape: BoxShape.circle,
-        ),
+        margin: EdgeInsets.only(right: widget.shape.spacing),
+        child: getNormalDotChildContainer(),
+        decoration: getNormalDotDecoration(color),
       );
     }
 
     if (widget.length % 2 == 1 && index > (widget.length ~/ 2)) {
       return Container(
-        margin: EdgeInsets.only(left: widget.spacing.toDouble()),
-        child: Container(
-          width: widget.size.toDouble(),
-          height: widget.size.toDouble(),
-        ),
-        decoration: BoxDecoration(
-          // color: Color(0xff35affc),
-          color: color ?? Color(0xff35affc),
-          shape: BoxShape.circle,
-        ),
+        margin: EdgeInsets.only(left: widget.shape.spacing),
+        child: getNormalDotChildContainer(),
+        decoration: getNormalDotDecoration(color),
       );
     }
 
     if ((widget.length % 2 == 0 && index < (widget.length ~/ 2)) ||
         index == -1) {
       return Container(
-        margin: EdgeInsets.only(right: widget.spacing.toDouble()),
-        child: Container(
-          width: widget.size.toDouble(),
-          height: widget.size.toDouble(),
-        ),
-        decoration: BoxDecoration(
-          color: color ?? Color(0xff35affc),
-          shape: BoxShape.circle,
-        ),
+        margin: EdgeInsets.only(right: widget.shape.spacing),
+        child: getNormalDotChildContainer(),
+        decoration: getNormalDotDecoration(color),
       );
     }
 
     if (widget.length % 2 == 0 && index > (widget.length ~/ 2)) {
       return Container(
-        margin: EdgeInsets.only(left: widget.spacing.toDouble()),
-        child: Container(
-          width: widget.size.toDouble(),
-          height: widget.size.toDouble(),
-        ),
-        decoration: BoxDecoration(
-          color: color ?? Color(0xff35affc),
-          shape: BoxShape.circle,
-        ),
+        margin: EdgeInsets.only(left: widget.shape.spacing),
+        child: getNormalDotChildContainer(),
+        decoration: getNormalDotDecoration(color),
       );
     }
 
     return Container(
-      child: Container(
-        width: widget.size.toDouble(),
-        height: widget.size.toDouble(),
-      ),
-      decoration: BoxDecoration(
-        color: color ?? Color(0xff808080),
-        shape: BoxShape.circle,
-      ),
+      child: getNormalDotChildContainer(),
+      decoration: getNormalDotDecoration(color),
     );
   }
 
   List<Widget> _renderNormalDots() {
     var listDots = List<Widget>();
     for (int i = 0; i < widget.length; i++) {
-      listDots.add(buildDot(widget.color ?? (0xff808080), i));
+      listDots.add(buildDot(widget.color ?? defaultNormalDotColor, i));
     }
     return listDots;
   }
@@ -134,8 +136,7 @@ class _DotsIndicatorState extends State<WormIndicator> {
             child: DotInstance(
               length: widget.length,
               listenable: widget.controller,
-              size: widget.size,
-              spacing: widget.spacing,
+              shape: widget.shape,
               color: widget.indicatorColor,
             ),
           ),
@@ -144,4 +145,3 @@ class _DotsIndicatorState extends State<WormIndicator> {
     );
   }
 }
-